Add Carthage compatibility - generate framework as dynamic
