+++ Cartfile
@@ -1,0 +1,1 @@
+binary "https://code.videolan.org/videolan/VLCKit/raw/master/Packaging/TVVLCKit.json"
