+++ Cartfile.resolved
@@ -1,0 +1,1 @@
+binary "https://code.videolan.org/videolan/VLCKit/raw/master/Packaging/TVVLCKit.json" "3.2.0"
