+++ Demo/Assets.xcassets/App Icon & Top Shelf Image.brandassets/Contents.json
