--- Demo/Demo/Assets.xcassets/LaunchImage.launchimage/Contents.json
@@ -1,15 +1,0 @@
-{
-  "images" : [
-    {
-      "orientation" : "landscape",
-      "idiom" : "tv",
-      "extent" : "full-screen",
-      "minimum-system-version" : "9.0",
-      "scale" : "1x"
-    }
-  ],
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
