--- Demo/Podfile
@@ -1,11 +1,0 @@
-platform :tvos, '11.0'
-
-source 'https://github.com/CocoaPods/Specs'
-
-target 'Demo' do
-  # Comment the next line if you're not using Swift and don't want to use dynamic frameworks
-  use_frameworks!
-
-  # Pods for TVVLCPlayer
-  pod 'TVVLCPlayer', :path => "../TVVLCPlayer.podspec"
-end
