+++ Modules/module.modulemap
@@ -1,0 +1,6 @@
+framework module TVVLCKit {
+  umbrella header "TVVLCKit.h"
+
+  export *
+  module * { export * }
+}
