+++ Sources/TVVLCPlayer-Bridging-Header.h
@@ -6,4 +6,4 @@
 //  Copyright © 2018 Jérémy Marchand. All rights reserved.
 //
 
-#import <TVVLCKit/TVVLCKit.h>
+//#import <TVVLCKit/TVVLCKit.h>
