+++ TVVLCPlayer/TVVLCPlayer.h
@@ -1,0 +1,20 @@
+//
+//  TVVLCPlayer.h
+//  TVVLCPlayer
+//
+//  Created by Jérémy Marchand on 18/08/2018.
+//  Copyright © 2018 Jérémy Marchand. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for TVVLCPlayer.
+FOUNDATION_EXPORT double TVVLCPlayerVersionNumber;
+
+//! Project version string for TVVLCPlayer.
+FOUNDATION_EXPORT const unsigned char TVVLCPlayerVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <TVVLCPlayer/PublicHeader.h>
+
+#import "TVVLCPlayer-Bridging-Header.h"
+//#import <TVVLCKit/TVVLCKit.h>
